+++ src/service/comment.ts
@@ -1,4 +1,6 @@
-import { Comment, isCommentCreate, isCommentUpdate } from '../domain/comment';
+import { Comment } from '../domain/comment/comment';
+import { CommentCreate, isCommentCreate } from '../domain/comment/commentCreate';
+import { CommentUpdate, isCommentUpdate } from '../domain/comment/commentUpdate';
 import commentRepository from '../repository/comment';
 
 const findById = async (id: number) => {
@@ -15,21 +17,31 @@ const findByCommunityId = async (communityId: number) => {
   return result;
 };
 
-const save = async (comment: Comment) => {
-  if (!isCommentCreate) return undefined;
-  const result = await commentRepository.save(comment);
+const findByUserId = async (communityId: number) => {
+  if (isNaN(communityId)) return undefined;
+  const result = await commentRepository.findByUserId(communityId);
   if (!result) return undefined;
   return result;
 };
 
-const update = async (id: number, comment: Comment) => {
-  if (typeof id !== 'number' || !isCommentUpdate) return undefined;
+const save = async (userId: number, comment: CommentCreate) => {
+  if (!isCommentCreate) return undefined;
+  const result = await commentRepository.save(userId, comment);
+  if (!result) return undefined;
+  return result;
+};
+
+const update = async (userId: number, id: number, comment: CommentUpdate) => {
+  const commentOrigin = await commentRepository.findById(id);
+  if (commentOrigin && commentOrigin.user_id != userId) return undefined;
   const result = await commentRepository.update(id, comment);
   if (!result) return undefined;
   return result;
 };
 
-const deleteById = async (id: number) => {
+const deleteById = async (userId: number, id: number) => {
+  const commentOrigin = await commentRepository.findById(id);
+  if (commentOrigin && commentOrigin.user_id != userId) return undefined;
   if (typeof id !== 'number') return undefined;
   const result = await commentRepository.deleteById(id);
   if (!result) return undefined;
@@ -42,6 +54,7 @@ const commentService = {
   save,
   update,
   deleteById,
+  findByUserId,
 };
 
 export default commentService;
